Add files via upload
A project I had in my mind and just stopped working on it for working reasons. The goal was to make a website for the union I was in with the following tools: a basic login, a place where they could keep track of how many days they worked and in which function they worked, with a preview of how much they earned working in the current week and in the current month. I plan to finish it some day, but for now I'll just keep it here as an example of what I've already learned